+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-7144] Introduce simple query string query for rules and implementations
+
+### Changed
+
+- [TD-7136] Merge old content with new one on file upload
+
+### Fixed
+
+- [TD-7136] Alias format from note in download
+
 ## [7.2.1] 2025-02-26
 
 ### Changed